// src/schemas/generated/tainacan.zod.ts
export const ItemEmbeddedMetadata = zod.object({
  "id": zod.number().optional(),
  "name": zod.string().optional(),
  "value": zod.unknown().optional(),
  "value_as_html": zod.string().optional(),
  "value_as_string": zod.string().optional(),
  "semantic_uri": zod.string().optional(),
  "multiple": zod.enum(['yes', 'no']).optional(),
  "mapping": zod.union([zod.looseObject({

}),zod.array(zod.unknown()),zod.string()]).optional(),
  "date_i18n": zod.string().optional(),
  "parent_meta_id": zod.string().optional()
}).describe('Embedded metadata value on an item (item.metadata[slug]). Distinct from metadatum definition.')

export type ItemEmbeddedMetadata = zod.input<typeof ItemEmbeddedMetadata>;
export type ItemEmbeddedMetadataOutput = zod.output<typeof ItemEmbeddedMetadata>;

export const Item = zod.object({
  "id": zod.number().optional().describe('Unique identifier'),
  "status": zod.string().optional().describe('WordPress post status'),
  "title": zod.string().optional().describe('Title of the item'),
  "description": zod.string().optional().describe('The item description'),
  "collection_id": zod.union([zod.number().describe('The collection ID'),zod.string().describe('The collection ID')]).optional().describe('The collection ID'),
  "author_id": zod.string().optional().describe('The item author\'s user ID (numeric string)'),
  "author_name": zod.string().optional().describe('Author display name'),
  "slug": zod.string().optional().describe('Item slug'),
  "creation_date": zod.string().optional().describe('Creation date'),
  "modification_date": zod.string().optional().describe('Modification date'),
  "terms": zod.array(zod.union([zod.string(),zod.number()])).nullish().describe('Term IDs'),
  "document_type": zod.enum(['attachment', 'url', 'text', 'empty']).optional().describe('Document type'),
  "document": zod.string().optional().describe('Document value'),
  "document_options": zod.union([zod.looseObject({

}).describe('Document display options'),zod.string().describe('Document display options')]).optional().describe('Document display options'),
  "document_content_index": zod.string().optional().describe('Document content for search'),
  "document_mimetype": zod.string().optional().describe('Document MIME type'),
  "document_as_html": zod.string().optional().describe('Rendered document HTML'),
  "_thumbnail_id": zod.union([zod.number().describe('Thumbnail attachment ID'),zod.string().describe('Thumbnail attachment ID')]).optional().describe('Thumbnail attachment ID'),
  "thumbnail": zod.looseObject({

}).optional().describe('Thumbnail URLs by size'),
  "thumbnail_alt": zod.string().optional().describe('Thumbnail alt text'),
  "thumbnail_id": zod.union([zod.number().describe('Thumbnail ID alias'),zod.string().describe('Thumbnail ID alias')]).optional().describe('Thumbnail ID alias'),
  "comment_status": zod.enum(['open', 'closed']).optional().describe('Comment status'),
  "url": zod.string().optional().describe('Public item URL'),
  "exposer_urls": zod.array(zod.looseObject({

})).optional().describe('Exposer URLs'),
  "supports_document_content_extraction": zod.boolean().optional().describe('Whether document content extraction is supported'),
  "metadata": zod.record(zod.string(), zod.object({
  "id": zod.number().optional(),
  "name": zod.string().optional(),
  "value": zod.unknown().optional(),
  "value_as_html": zod.string().optional(),
  "value_as_string": zod.string().optional(),
  "semantic_uri": zod.string().optional(),
  "multiple": zod.enum(['yes', 'no']).optional(),
  "mapping": zod.union([zod.looseObject({

}),zod.array(zod.unknown()),zod.string()]).optional(),
  "date_i18n": zod.string().optional(),
  "parent_meta_id": zod.string().optional()
}).describe('Embedded metadata value on an item (item.metadata[slug]). Distinct from metadatum definition.')).optional().describe('Metadata values keyed by metadatum slug'),
  "current_user_can_edit": zod.boolean().optional(),
  "current_user_can_delete": zod.boolean().optional()
}).describe('Tainacan item (GET \/items\/{id}). Patched from REST controller prepare_item_for_response.')

export type Item = zod.input<typeof Item>;
// ...
export const Term = zod.object({
  "id": zod.number().optional().describe('Unique identifier for the object.'),
  "status": zod.string().optional().describe('Status for control of visibility and access.'),
  "term_id": zod.number().optional().describe('Unique identifier'),
  "name": zod.string().optional().describe('Name of the term'),
  "parent": zod.union([zod.number(),zod.string()]).nullish(),
  "description": zod.string().optional().describe('The term description'),
  "taxonomy": zod.string().optional().describe('The term taxonomy'),
  "user": zod.union([zod.number(),zod.string()]).optional(),
  "header_image_id": zod.union([zod.number(),zod.string()]).nullish(),
  "cover_page_id": zod.union([zod.number(),zod.string()]).nullish(),
  "hide_empty": zod.boolean().nullish(),
  "current_user_can_edit": zod.boolean().optional().describe('Whether current user can edit this object'),
  "current_user_can_delete": zod.boolean().optional().describe('Whether current user can delete this object'),
  "url": zod.string().optional(),
  "thumbnail": zod.union([zod.looseObject({

}),zod.string(),zod.boolean()]).optional(),
  "thumbnail_alt": zod.union([zod.string(),zod.boolean()]).optional(),
  "thumbnail_id": zod.union([zod.number(),zod.string()]).optional(),
  "header_image": zod.union([zod.looseObject({

}),zod.string(),zod.boolean()]).optional(),
  "hierarchy_path": zod.string().optional(),
  "total_children": zod.union([zod.number(),zod.string()]).optional()
})

export type Term = zod.input<typeof Term>;
// ...
export const FilterMetadatum = zod.object({
  "metadatum_id": zod.union([zod.number(),zod.string()]).optional(),
  "metadatum_name": zod.string().optional(),
  "metadata_type_object": zod.object({
  "primitive_type": zod.string().optional(),
  "options": zod.union([zod.object({
  "taxonomy_id": zod.number().optional(),
  "taxonomy": zod.string().optional()
}),zod.array(zod.unknown())]).optional()
}).optional()
}).describe('Nested metadatum summary on a filter (Filter._toArray).')

export type FilterMetadatum = zod.input<typeof FilterMetadatum>;
export type FilterMetadatumOutput = zod.output<typeof FilterMetadatum>;

export const Filter = zod.object({
  "id": zod.number().optional().describe('Unique identifier'),
  "status": zod.string().optional().describe('Status for control of visibility and access.'),
  "name": zod.string().optional().describe('Name of the filter'),
  "order": zod.union([zod.string().describe('Filter order. This metadata is used if filters were manually ordered.'),zod.number().describe('Filter order. This metadata is used if filters were manually ordered.')]).optional().describe('Filter order. This metadata is used if filters were manually ordered.'),
  "description": zod.string().optional().describe('The filter description'),
  "filter_type_options": zod.union([zod.array(zod.unknown()).describe('The filter type options'),zod.looseObject({

}).describe('The filter type options'),zod.string().describe('The filter type options')]).optional().describe('The filter type options'),
  "filter_type": zod.string().optional().describe('The filter type class name, such as filter_type: Tainacan\\Filter_Types\\Checkbox'),
  "begin_with_filter_collapsed": zod.string().optional().describe('With this option enabled, the filter will appear as a button with an add icon, that should be pressed prior to loading any facet information.'),
  "display_in_repository_level_lists": zod.string().optional().describe('With this option enabled, the filter will appear even in repository level items lists, such as the complete items list and the term items list.'),
  "collection_id": zod.union([zod.number(),zod.string()]).optional(),
  "metadatum_id": zod.union([zod.number(),zod.string()]).optional(),
  "max_options": zod.union([zod.number().describe('The maximum number of options to be loaded by default on the filter.'),zod.string().describe('The maximum number of options to be loaded by default on the filter.')]).optional().describe('The maximum number of options to be loaded by default on the filter.'),
  "placeholder": zod.string().optional().describe('The filter input placeholder. This is a simple message that will appear inside textual input and may indicate to the user what kind of information is expected.'),
  "description_bellow_name": zod.string().optional().describe('Whether the filter description should be displayed below the input label instead of inside a tooltip.'),
  "current_user_can_edit": zod.boolean().optional().describe('Whether current user can edit this object'),
  "current_user_can_delete": zod.boolean().optional().describe('Whether current user can delete this object'),
  "metadatum": zod.object({
  "metadatum_id": zod.union([zod.number(),zod.string()]).optional(),
  "metadatum_name": zod.string().optional(),
  "metadata_type_object": zod.object({
  "primitive_type": zod.string().optional(),
  "options": zod.union([zod.object({
  "taxonomy_id": zod.number().optional(),
  "taxonomy": zod.string().optional()
}),zod.array(zod.unknown())]).optional()
}).optional()
}).optional().describe('Nested metadatum summary on a filter (Filter._toArray).'),
  "filter_type_object": zod.looseObject({

}).optional(),
  "inherited": zod.boolean().optional(),
  "enabled": zod.enum(['yes', 'no']).optional()
})

export type Filter = zod.input<typeof Filter>;
// ...
export const Taxonomy = zod.object({
  "id": zod.number().optional().describe('Unique identifier'),
  "status": zod.string().optional().describe('Status for control of visibility and access.'),
  "name": zod.string().optional().describe('Name of the taxonomy'),
  "description": zod.string().optional().describe('The taxonomy description'),
  "slug": zod.string().optional().describe('The taxonomy slug'),
  "allow_insert": zod.string().optional().describe('Allow\/Deny the creation of new terms in the taxonomy'),
  "hierarchical": zod.string().optional().describe('Allow\/Deny the existence of terms children to build a hierarchy'),
  "enabled_post_types": zod.union([zod.string(),zod.array(zod.unknown())]).optional(),
  "collections_ids": zod.union([zod.string(),zod.array(zod.unknown()),zod.number()]).optional(),
  "current_user_can_edit": zod.boolean().optional().describe('Whether current user can edit this object'),
  "current_user_can_delete": zod.boolean().optional().describe('Whether current user can delete this object')
})

export type Taxonomy = zod.input<typeof Taxonomy>;
// ...
export type Collection = zod.input<typeof Collection>;